LayoutTests/pointer-lock/pointerlocklost-event-expected.txt

Test that pointerlocklost event is dispatched.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.layoutTestController is defined.
PASS navigator.webkitPointer.isLocked is true
JavaScript initiated unlock.
PASS webkitpointerlocklost handler call 0
PASS webkitpointerlocklost handler call 1
PASS All expected webkitpointerlocklost events received.
PASS navigator.webkitPointer.isLocked is false
PASS navigator.webkitPointer.isLocked is true
Host initiated unlock.
PASS webkitpointerlocklost handler call 0
PASS webkitpointerlocklost handler call 1
PASS All expected webkitpointerlocklost events received.
PASS navigator.webkitPointer.isLocked is false
PASS successfullyParsed is true

TEST COMPLETE
